Finial Commit
Ready for review. Renamed all of the files to remove the _txt. Renamed the homepage to index.html so it would work with github pages. Formatted the grids by clearing the floats and expanding to rows to cover the content. The column's width was also set. The color formatting of the page including the header and background was also done. Move the footer to the correct area in the bottom right corner. Set the font family for the webpage. Finally, the width of the body was change to 95%.